+++ mc_web_console_new_front/src/shared/libs/api/instance.ts
@@ -1,6 +1,7 @@
 import axios from 'axios';
 
-const url = 'http://mcmpdemo.csesmzc.com:3000';
+// const url = 'http://mcmpdemo.csesmzc.com:3000';
+const url = 'http://localhost:3000';
 
 const createInstance = () => {
   return axios.create({
